fix: 빌드 에러 수정 Type error: Route src/app/api/auth/[...nextauth]/route.ts does not match the required types of a Next.js Route.   auth is not a valid Route export field.
