+++ src/app/api/poll/[id]/route.ts
@@ -1,6 +1,6 @@
 import {deletePoll, fetchDetailPollItem, postPollOption} from '@/db/poll';
 import {NextRequest, NextResponse} from 'next/server';
-import {auth} from '../../auth/[...nextauth]/route';
+import {auth} from '../../auth/[...nextauth]/auth.util';
 
 export async function GET(request: NextRequest, {params: {id}}: {params: {id: string}}) {
   const session = (await auth()) as any;
